+++ src/main/java/app/habitzl/elasticsearch/status/monitor/tool/configuration/CliOptions.java
@@ -19,6 +19,7 @@ public class CliOptions {
     protected static final String USER_OPTION_LONG = "username";
     protected static final String PASSWORD_OPTION_LONG = "password";
     protected static final String REPORT_FILES_PATH_OPTION_LONG = "reportPath";
+    protected static final String SKIP_ARCHIVE_REPORT_LONG = "skipArchiveReport";
 
     private final Options availableOptions;
 
@@ -41,6 +42,7 @@ public class CliOptions {
         options.addOption(createUserOption());
         options.addOption(createPasswordOption());
         options.addOption(createReportFilesPathOption());
+        options.addOption(createSkipArchiveReportOption());
         return options;
     }
 
@@ -120,6 +122,14 @@ public class CliOptions {
                 .hasArg(true)
                 .argName("report files path")
                 .desc("The path to the location of the generated report files.")
+                .build();
+    }
+
+    private Option createSkipArchiveReportOption() {
+        return Option.builder()
+                .longOpt(SKIP_ARCHIVE_REPORT_LONG)
+                .hasArg(false)
+                .desc("Skips the generation of an archive report file located in a timestamp folder.")
                 .build();
     }
 }